Änderungen an der Oberfläsche

+++ src/main/java/com/fhdw/loeppe/dto/Customer.java
@@ -4,12 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.validation.constraints.NotEmpty;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 public class Customer {
-
-    private long id;
+    private int id;
     private String firstname;
     private String lastname;
     private String address;
